requirements.txt
anyio==4.12.0
APScheduler==3.11.2
certifi==2025.11.12
filelock==3.20.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
python-dotenv==1.2.1
python-telegram-bot==22.5
tzdata==2025.3
tzlocal==5.3.1
ujson==5.11.0
